Save config file occasionally and after UI update
The config file save can take a long time, especially for large
models. This commit makes sure that it doesn't happen after every
action, and that it happens after the new info is sent to the UI
so that there is no lag.

Currently, the config file will be saved every 2 seconds if there
are changes, which is fast enough that information should rarely
be lost because of e.g. a crash, but slow enough that the UI
is considerably more responsive.