Added info modal for nodes
This plots the output function of the node, if it's a simple function
of time. Of course, if it's a non-simple function of time it'll
still try, so it's a bit dangerous for things that communicate
with the external world, for example.

This also involved significant refactoring of viz_modal.js, in order
to not copy-paste the entirety of ensemble_body, since node_body is
almost identical. This might not be true for other modals, though.

Since I was adding in tooltip stuff, I also cleaned up those files
so that they're not just one big long line. In Javascript, property
access (obj.thing) is nearly identical to dictionary access
(obj["thing"]) just shorter, so used that where possible.