Added info modal for networks
This displays some statistics about the network (mostly how much
of whatever is in there) and the connection info.

In order to implement this, I made the `info_body` functions more
generic, since Networks have fewer tabs with different titles.

I also added a warning if there are no connections into an object;
without this, the tab is just disappointingly empty. This warning
says something about the network being expanded if it's a network
that is expanded (since expanded networks don't have connection info).