Fixed bug in reload re:iteritems
The switchfrom .items() to nengo.utils.compat.iteritems() wasn't
tested well enough here.  The loop changes the dictionary, and
iteritems() really doesn't like that.  This wasn't a problem with
the .items() approach, since it's only ever changing the item
it's on in the loop.  The fix here is to duplicate the dictionary.