Add methods to enable automatic port selection.
Setting the port to 0 will instruct the OS to chose a free port. But as
the default start() method directly enters serve_forever, there was no
way to get back the chosen port. This commit adds method for a two step
startup where first the server is created (which allows to get the
port) and then with the second call the serve_forever is entered.