Make Modal a class instead of singleton
While we won't do this, it makes it possible to make multiple modals.
The major advantage here is just that the Javascript code looks more
idiomatic.

The main reason I did this was because info_body was being used to
clear the body, which is overkill (and adds divs we don't want). So I
also added a clear_body function. And a text_body function, which the
reset modal now uses.