Add a 'script directory' configuration option.
This shows up in the main Configuration Options modal dialog accessible
through the gear at the top left. Right now, we expect you to type in
a full path to a directory, which isn't super friendly, but it works.

The built-in examples always show up. This required some changes to
`server.py`, but the logic should be the same if the script directory
is not changed.

This option is also managed through `Nengo.config`, which means that
it persists after shutting down and restarting the server.
This should reduce the amount of searching for files if you have
all of your scripts in a certain location.