Enable sliders on nodes with process outputs
This case was previously not considered and failed when a slider was
created. Now, sliders can be made on nodes with process outputs.

This commit also changes how node.output is overridden for all other
node output types. The changed version uses processes, so this change
breaks compatibility with Nengo 2.0.x, so that should be kept in mind
when determining when to merge this commit.

The new OverriddenOutput should be slightly faster in some cases because
the (former) `data` list is now a numpy array where `np.nan` indicates
that the value should not be overridden. This array is only modified,
so we save several instances of making a new list of Nones, which may
make a marginal difference in some cases.

Another possible optimization is that data is only sent to the client
to update the slider position when a value has changed. Previously,
data was sent to the client on every timestep of a callable output.
Additionally, data is now sent to the client for all situations,
not just callables.