Conform version to PEP440 and simplify setup.py
PEP440 (https://www.python.org/dev/peps/pep-0440/) defines a standard
for Python versions that setuptools adheres to. It doesn't allow
hyphens for marking as a dev version, so we switch to using a dot
as the standard requires.

The main benefit here is to avoid a warning given by setuptools
when running `setup.py` on a dev version.

This commit also contains some minor improvements to setup.py,
mostly changing the author information and having fewer variables.