Add 'nengo' script, deprecate 'nengo_gui'
I did the nengo_gui deprecation with regular print statements rather
than warnings, as warnings can be turned off, and are pretty difficult
to read in the terminal, which is where users will see this warning.
The server already prints things to the terminal, so it's not unusual
for the server to do this.