Reworks the toolbar, adding a side navigation bar
The file open menu is now done with the sidebar, rather than in its own
window (and thus no longer disappears if your mouse cursor goes outside
it). Many of the options in the top toolbar have been moved to the side
bar, clearing up the top menu a bit.

Co-authored-by: Terry Stewart <tcstewar@uwaterloo.ca>