Use CSS instead of <em> for built-in examples
Using the <em> tag caused an issue with newer versions of
jqueryfiletree because they look at the direct parent of the
element that got a click event; since that element was the <em>
tag, the parent was <a>, but jqueryfiletree expected that to be
the <li> element, and therefore could not determine if the <li>
corresponded to a file or directory.

Fixed this by removing the <em> tag from the built-in examples
directory entry in the file tree, and instead use a CSS rule to
make that link italicized (but not any descendants).