Start nengo_gui server as fixture for tests.
This allows to run all tests with a simple `py.test nengo_gui`
without the need to start the server manually. It also automatically
selects a free port allowing to run another nengo_gui server or other
things on the default things in parallel (e.g. to work on an actual
model while the tests run).