Convert JavaScript files to TypeScript.
The goal for this commit was to get all of the JavaScript files
converted to TypeScript with minimal changes, as was done for the
modularization commit.

While much of this was straightforward (rename .js to .ts, change
require statements to import and change export style), I ran into more
issues than I was expecting.

TypeScript really didn't like our prototype-based objects, and I could
not figure out how to get them working without converting them to
classes, so that's what I did. However, I attempted to do this with
as few lines changed as possible, so I kept the current indentation
level of all of the functions even though they are now wrapped
within the class declaration (I'll do all the reindenting in another
commit).

While type definitions for most of our dependencies were either easy to
find, or not necessary to get the GUI to compile and run properly, the
definitions for the Ace editor did not work because we import through
brace. To deal with this, I wrote some code to convert the Ace
definitions to work with Brace. Once that code is merged in Brace,
this commit will work without modifications as the definition file will
be accessible through npm.

The TypeScript compiler found several subtle JS bugs that had to be
fixed in order to compile.

- In several instances we were not the 'var' keyword where we should
  have. Withou 'var', the variable leaks out to the global scope,
  so everything works 99.9% of the time, but that 0.1% sucks to debug.
- In several instances, we were referring to what seems like a
  non-existent variable. We either never hit these code path, or
  the variable was in the global scope (possibly intentionally,
  possibly through forgetting var).
- We used the deprecated global JS function escape. Replaced with
  encodeURIComponent.