Refactor debug interface
The biggest change is to make the main canvas an iframe rather than
a div. This allows modals to be displayed but still interacted with
through the controls at the bottom.

This required making multiple entry points in the webpack
configuration, so it seemed a good time to upgrade to Webpack 2.