Require token based authentication.
This resolves some security problems on multi-user systems. Prior to
these changes any user on the local system could have connected to the
Nengo GUI server to run arbitrary code under the permissions of
a different user. With the token based authentication such an evil user
has to obtain the token first.

This is similar to how Jupyter Notebook handles things.

Fixes #867.