Update setup.py and MANIFEST.in
It's no longer recommended to include ez_setup.py; setuptools
should be available in the overwhelming majority of cases,
and if it's not available it's likely the user will have to
update their Python install anyhow.

Also updates the MANIFEST file to ensure that all files are
included in the distribution (as tested by `check-manifest`).