Do not persist sessions for non-auth users.
Merely connecting to the server used to create a new session and store
it in memory indefinitely (sessions are only deleted if they are
retrieved again after their expiry). This would allow a malicious user
to consume a large amount of memory by issuing a large number of
requests. With this commit a session will be discarded after a request
as long as a user is not authenticated.

Fixes #914.