Use one-time token to open browser.
Opening the browser will expose the token shortly through the system's
process list which allows another user on the system to steal the token
to login and run code under a different user's permissions. By limiting
this particular token to a one-time use, the attacker would need to
steal the token and use it before the automatically invoked browser
requested the page.

Jupyter notebook also generates a one-time token for the automatically
launched browser.