Untie password from listening on ext. interfaces.
Listening on external interfaces still requires SSL (to prevent
unencrypted tokens or passwords from being stolen), but we do not
enforce a password because people usually choose weak passwords and
the token authentication is probably more secure. This also allows
to set a password on a server that is not listening for external
connections for convenience (in case someone does not want to copy
tokens).