Use editor from page settings (ignore config)
Relying on the config can cause problems. For instance, when the config
was writting by the standard GUI with ACE editor, it will load the ACE
editor also when reusing that config within IPythonViz. But the editor
needs to be deactivated in IPythonViz to have it work properly (the
editor sends empty code that gets compiled and is then used instead of
the actual model passed to IPythonViz).

Fixes #846.