Set prefix based on first request if None.
There seems to be no better way of obtaining the correct prefix when
executed from a Jupyter notebook. The InlineGUI has no access to the
notebooks base_url. For security reasons, the prefix can only be set
when it is None (which it should only be for InlineGUI) and if a
valid authentication token is provided. Because the prefix has to be
set before normal request processing and authentication is done (both
rely on the prefix), this is done in manually in http_GET including
checking the token.