Prevent failure for SPA Transcode plots.
Fixes #970.

This is only a hotfix preventing a complete crash of the GUI. The
functionality to override the output value for Transcode modules is
still not working with this even though the context menu suggests
otherwise. In the future, values should be overwritten in the same
way as we do for nodes.