Reset pan and zoom on autolayout
Previously, when applying an autolayout, the object coordinates were
rescaled such that they fit the current viewport taking pan and zoom
into account.

Unfortunately, this causes (a) numerical instability due to exponentially
exploding/imploding coordinates, (b) issues with scale-dependent objects
such as the multi-connection arcs from this patch set.

This commit resets pan and zoom instead of rescaling the coordinates
in a new "feedforward_layout_done" message handler. Since this would
also change the coordinates of the non-netgraph components, the
transformation inverse to the pan/zoom reset is applied to these
components.